fix: E1000 driver hangs due to wrong MMIO access method
- Change e1000_io_base from uint16_t to volatile uint32_t* for MMIO
- Replace outd/ind port I/O with direct memory pointer dereferencing
- Fix e1000_init() parameter type from uint16_t to uint32_t
- Fix NULL check in e1000_send()

+++ include/net/e1000.h
@@ -74,7 +74,7 @@ typedef struct e1000_rx_desc {
 
 extern uint8_t e1000_mac[6];
 
-int e1000_init(uint16_t io_base);
+int e1000_init(uint32_t io_base);
 int e1000_send(uint8_t* data, size_t len);
 void e1000_receive();
 uint8_t* e1000_arp_cache_lookup(uint32_t ip);
